# src/main/java/src/main/dto/auth/AuthResponse.java
package src.main.dto.auth;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthResponse {
    
    private String token;
    private String email;
    private String name;
    private boolean ai_access_enabled;
} 