# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MoneyGuard API
  description: |
    # MoneyGuard API
    
    API для приложения управления личными финансами MoneyGuard, которое позволяет пользователям отслеживать доходы и расходы, 
    устанавливать бюджеты, создавать финансовые цели и получать аналитику по своим финансам.
    
    ## Аутентификация
    
    Все защищенные эндпоинты требуют JWT токен, который можно получить через `/auth/login` или `/auth/register`.
    Токен необходимо передавать в заголовке `Authorization` в формате `Bearer {token}`.
    
    ## Доступ к ИИ
    
    Приложение предоставляет базовый функционал для всех пользователей. Доступ к AI-рекомендациям можно получить,
    активировав соответствующую опцию через API подписки.
  version: 1.0.0
  contact:
    email: 123
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации, полученный при регистрации или входе в систему
...
  schemas:
    RegisterRequest:
      type: object
      required: [ email, password, name ]
      properties:
        email:
          type: string
          format: email
          description: Email пользователя (должен быть уникальным)
          example: user@example.com
        password:
          type: string
          minLength: 8
          description: Пароль пользователя (минимум 8 символов,заглавная, прописная, цифра)
          example: password123
        name:
          type: string
          description: Имя пользователя
          example: Иван Иванов
    LoginRequest:
      type: object
      required: [ email, password ]
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@example.com
        password:
          type: string
          description: Пароль пользователя
          example: password123
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT токен для авторизации
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        email:
          type: string
          description: Email пользователя
          example: user@example.com
        name:
          type: string
          description: Имя пользователя
          example: Иван Иванов
        ai_access_enabled:
          type: boolean
          description: Флаг доступа к AI-рекомендациям
          example: false
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP статус код ошибки
          example: 404
        message:
          type: string
          description: Сообщение об ошибке
          example: Категория с ID 999 не найдена
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2023-04-01T12:00:00Z
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: /api/categories/999
    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP статус код ошибки
          example: 400
        message:
          type: string
          description: Общее сообщение об ошибке
          example: Ошибка валидации данных
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2023-04-01T12:00:00Z
        path:
          type: string
          description: Путь запроса, вызвавшего ошибку
          example: /api/categories
        errors:
          type: object
          additionalProperties:
            type: string
          description: Детальные ошибки валидации по полям
          example:
            name: Название категории не может быть пустым
            icon: Иконка должна быть указана
paths:
  /auth/register:
    post:
      tags: [ Authentication ]
      summary: Регистрация нового пользователя
      description: |
        Регистрирует нового пользователя в системе и возвращает JWT токен для аутентификации.
        
        Требования к данным:
        - Email должен быть уникальным и соответствовать формату email
        - Пароль должен содержать минимум 8 символов, заглавную, прописную, цифру
        - Имя пользователя не должно быть пустым
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              example1:
                summary: Пример запроса на регистрацию
                value:
                  email: user@example.com
                  password: Password123
                  name: Иван Иванов
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                example1:
                  summary: Пример успешного ответа
                  value:
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    email: user@example.com
                    name: Иван Иванов
                    ai_access_enabled: false
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                example1:
                  summary: Пример ошибки валидации
                  value:
                    status: 400
                    message: Ошибка валидации данных
                    timestamp: 2023-04-01T12:00:00Z
                    path: /auth/register
                    errors:
                      email: Email должен быть корректным
                      password: Пароль должен содержать минимум 6 символов
        '409':
          description: Пользователь с таким email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example1:
                  summary: Пример ошибки дублирования
                  value:
                    status: 409
                    message: Пользователь с email user@example.com уже существует
                    timestamp: 2023-04-01T12:00:00Z
                    path: /auth/register
  /auth/login:
    post:
      tags: [ Authentication ]
      summary: Аутентификация пользователя
      description: |
        Аутентифицирует пользователя по email и паролю, возвращает JWT токен для дальнейшей авторизации.
        
        В случае успешной аутентификации возвращается токен, который необходимо использовать в заголовке 
        Authorization для доступа к защищенным ресурсам.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              example1:
                summary: Пример запроса на вход
                value:
                  email: user@example.com
                  password: password123
      responses:
        '200':
          description: Пользователь успешно аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                example1:
                  summary: Пример успешного ответа
                  value:
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    email: user@example.com
                    name: Иван Иванов
                    ai_access_enabled: false
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example1:
                  summary: Пример ошибки аутентификации
                  value:
                    status: 401
                    message: Неверный email или пароль
                    timestamp: 2023-04-01T12:00:00Z
                    path: /auth/login
  /auth/logout:
    post:
      tags: [ Authentication ]
      summary: Выход из системы
      description: |
        Выполняет выход пользователя из системы, инвалидирует текущий JWT токен.
        
        После успешного выхода текущий токен становится недействительным и не может быть использован 
        для доступа к защищенным ресурсам.
      security:
        - bearerAuth: [ ]
      responses:
        '204':
          description: Пользователь успешно вышел из системы
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...